Fix warning about char indexing

+++ src/controls/code_edit.cpp
@@ -264,7 +264,7 @@ void Lexer::lex(Slice<const char> source) {
             tokens.data[tokens.length++] = Token(Token::Type::Number, pos.index);
             pos.next();
             while (pos.index < source.length) {
-                if (!is_number[source.data[pos.index]]) {
+                if (!is_number[(u8)source.data[pos.index]]) {
                     break;
                 }
                 pos.next();
